refactor: move read guards to guards.py and restrict allowlist to MANAGED only
