+++ tests/adapters/databricks/sql/test_rows.py
@@ -13,15 +13,12 @@ from pyspark.sql import Row
 import pytest
 
 from delta_engine.adapters.databricks.sql.rows import (
-    UnsupportedCatalogRelationError,
     clustering_columns_from_detail_row,
     column_tags_from_rows,
     foreign_keys_from_rows,
     managed_properties_from_detail_row,
     primary_key_from_rows,
     referencing_foreign_keys_from_rows,
-    require_delta_format,
-    require_supported_relation,
     table_tags_from_rows,
 )
 from delta_engine.domain.model import ForeignKeyReference, QualifiedName
@@ -231,50 +228,3 @@ def test_detail_clustering_absent_field_or_empty_array_means_unclustered():
     assert clustering_columns_from_detail_row(Row(properties={})) == ()
     assert clustering_columns_from_detail_row(Row(clusteringColumns=[])) == ()
     assert clustering_columns_from_detail_row(Row(clusteringColumns="[]")) == ()
-
-
-QN = QualifiedName("cat", "sch", "tbl")
-
-
-# ---------- relation-kind guard ----------
-
-
-@pytest.mark.parametrize("table_type", ["MANAGED", "EXTERNAL", "managed", "external"])
-def test_require_supported_relation_admits_ordinary_delta_tables(table_type):
-    require_supported_relation(table_type, QN)  # does not raise
-
-
-@pytest.mark.parametrize(
-    "table_type",
-    [
-        "VIEW",
-        "MATERIALIZED_VIEW",
-        "STREAMING_TABLE",
-        "FOREIGN",
-        "MANAGED_SHALLOW_CLONE",
-        "EXTERNAL_SHALLOW_CLONE",
-        "SOME_FUTURE_KIND",
-    ],
-)
-def test_require_supported_relation_rejects_every_other_kind(table_type):
-    with pytest.raises(UnsupportedCatalogRelationError):
-        require_supported_relation(table_type, QN)
-
-
-def test_require_supported_relation_names_the_object_and_kind():
-    with pytest.raises(UnsupportedCatalogRelationError, match="STREAMING_TABLE"):
-        require_supported_relation("STREAMING_TABLE", QN)
-
-
-# ---------- format guard ----------
-
-
-@pytest.mark.parametrize("table_format", ["delta", "DELTA", "Delta"])
-def test_require_delta_format_admits_delta(table_format):
-    require_delta_format(Row(format=table_format), QN)  # does not raise
-
-
-@pytest.mark.parametrize("table_format", ["iceberg", "parquet", "csv"])
-def test_require_delta_format_rejects_non_delta(table_format):
-    with pytest.raises(UnsupportedCatalogRelationError):
-        require_delta_format(Row(format=table_format), QN)
